fix build derp
strlen has a return type of size_t

+++ include/private/android_filesystem_config.h
@@ -336,7 +336,7 @@ static inline void fs_config(const char *path, int dir,
                              unsigned *uid, unsigned *gid, unsigned *mode, uint64_t *capabilities)
 {
     const struct fs_path_config *pc;
-    int plen;
+    size_t plen;
 
     if (path[0] == '/') {
         path++;
@@ -345,7 +345,7 @@ static inline void fs_config(const char *path, int dir,
     pc = dir ? android_dirs : android_files;
     plen = strlen(path);
     for(; pc->prefix; pc++){
-        int len = strlen(pc->prefix);
+        size_t len = strlen(pc->prefix);
         if (dir) {
             if(plen < len) continue;
             if(!strncmp(pc->prefix, path, len)) break;
